+++ .github/workflows/beta-publisher.yml
@@ -1,0 +1,24 @@
+name: Auto Publisher
+
+on:
+  push:
+    branches:
+      - releases/*
+
+jobs:
+  build-and-publish:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/setup-node@v1
+        with:
+          node-version: 10
+          registry-url: https://registry.npmjs.org/
+      - run: npm i
+      - run: npm i -g vsce
+      - run: npm run setup
+      - run: npm run publish-beta
+        env:
+          VSCE_TOKEN: ${{secrets.VSCE_TOKEN}}
+          NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
+          CLIENT_TOKEN: ${{ secrets.DEF_CLIENT_TOKEN }}
