+++ package.json
@@ -14,8 +14,11 @@
     "packages:build": "npm run packages:clean && ts-node ./scripts/build.ts",
     "packages:clean": "rm -rf ./packages/*/lib",
     "publish": "npm run publish:package && npm run publish:extension",
-    "publish:package": "ts-node ./scripts/package-publish.ts",
-    "publish:extension": "ts-node ./scripts/extension-publish.ts",
+    "publish:package": "ts-node ./scripts/publish-package.ts",
+    "publish:extension": "ts-node ./scripts/publish-extension.ts",
+    "publish-beta": "npm run publish-beta:package && npm run publish-beta:extension",
+    "publish-beta:package": "ts-node ./scripts/publish-beta-package.ts",
+    "publish-beta:extension": "ts-node ./scripts/publish-beta-extension.ts",
     "lint": "eslint --cache --ext .ts,.tsx,.js ./ --quiet",
     "lint:fix": "eslint --cache --ext .ts,.tsx,.js ./ --quiet --fix",
     "coverage": "codecov",
