+++ packages/vscode-webview/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iceworks/vscode-webview",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "description": "Webview related API for VSCode extension.",
   "files": [
     "lib"
