+++ scripts/getExtensionInfos.ts
@@ -1,16 +1,13 @@
-/**
- * Scripts to check unpublished version and run publish
- */
+
 import { existsSync, readdirSync, readFileSync } from 'fs';
 import { join } from 'path';
-import { spawnSync } from 'child_process';
 import axios from 'axios';
 
 const RETRY_LIMIT = 3;
 const TIMEOUT = 5000; // ms
 const TARGET_DIRECTORY = join(__dirname, '../extensions');
 
-interface IExtensionInfo {
+export interface IExtensionInfo {
   name: string;
   directory: string;
   localVersion: string;
@@ -34,28 +31,7 @@ function checkVersionExists(extension: string, version: string, retry = 0): Prom
     });
 }
 
-function publish(extension: string, directory: string, version: string): void {
-  // npm install
-  spawnSync('npm', [
-    'install',
-  ], {
-    stdio: 'inherit',
-    cwd: directory,
-  });
-
-  // vsce publish
-  console.log('[VSCE] PUBLISH: ', `${extension}@${version}`);
-  spawnSync('vsce', [
-    'publish',
-    '-p',
-    process.env.VSCE_TOKEN
-  ], {
-    stdio: 'inherit',
-    cwd: directory,
-  });
-}
-
-async function getExtensionInfos(): Promise<IExtensionInfo[]> {
+export async function getExtensionInfos(): Promise<IExtensionInfo[]> {
   const extensionInfos: IExtensionInfo[] = [];
 
   if (!existsSync(TARGET_DIRECTORY)) {
@@ -96,22 +72,3 @@ async function getExtensionInfos(): Promise<IExtensionInfo[]> {
   }
   return extensionInfos;
 }
-
-// Entry
-getExtensionInfos().then((extensionInfos: IExtensionInfo[]) => {
-  // Publish
-  console.log('');
-  if (extensionInfos.length === 0) {
-    console.log('[PUBLISH] No diff with all extensions.');
-  } else {
-    console.log('[PUBLISH] Will publish following extensions:');
-  }
-
-  for (let i = 0; i < extensionInfos.length; i++) {
-    const { name, directory, localVersion, shouldPublish } = extensionInfos[i];
-    if (shouldPublish) {
-      console.log(`--- ${name}@${localVersion} ---`);
-      publish(name, directory, localVersion);
-    }
-  }
-});
