+++ scripts/publish-beta-extension.ts
@@ -1,0 +1,43 @@
+/**
+ * Scripts to check unpublished version and run beta publish  
+ */
+import { spawnSync } from 'child_process';
+import { IExtensionInfo, getExtensionInfos } from './getExtensionInfos';
+
+function publish(extension: string, directory: string, version: string): void {
+  // npm install
+  spawnSync('npm', [
+    'install',
+  ], {
+    stdio: 'inherit',
+    cwd: directory,
+  });
+
+  // vsce package
+  console.log('[VSCE] PUBLISH: ', `${extension}@${version}`);
+  spawnSync('vsce', [
+    'package',
+  ], {
+    stdio: 'inherit',
+    cwd: directory,
+  });
+
+  // Upload to oss
+     
+}
+
+// Entry
+console.log('[PUBLISH BETA] Start:');
+getExtensionInfos().then((extensionInfos: IExtensionInfo[]) => {
+  // Publish
+  let publishedCount = 0;
+  for (let i = 0; i < extensionInfos.length; i++) {
+    const { name, directory, localVersion, shouldPublish } = extensionInfos[i];
+    if (shouldPublish) {
+      publishedCount++;
+      console.log(`--- ${name}@${localVersion} ---`);
+      publish(name, directory, localVersion);
+    }
+  }
+  console.log(`[PUBLISH BETA] Complete (count=${publishedCount}).`);
+});
